style: forge fmt
Apply `forge fmt` (default and test profiles) to files with pre-existing
formatting drift surfaced while running the formatter over the disable
feature, plus test-profile normalization in EigenPodUnit.t.sol. No
behavioral changes.

+++ src/test/unit/EigenPodUnit.t.sol
@@ -1958,11 +1958,7 @@ contract EigenPodUnitTests_DisableRestaking is EigenPodUnitTests {
         eigenPodManagerMock.setBeaconChainETHStrategy(BEACON_ETH_STRATEGY);
     }
 
-    function _newQueuedWithdrawal(address staker, uint32 startBlock)
-        internal
-        pure
-        returns (IDelegationManagerTypes.Withdrawal memory)
-    {
+    function _newQueuedWithdrawal(address staker, uint32 startBlock) internal pure returns (IDelegationManagerTypes.Withdrawal memory) {
         return _newQueuedWithdrawal(staker, startBlock, address(0));
     }
 
@@ -2221,9 +2217,7 @@ contract EigenPodUnitTests_DisableRestaking is EigenPodUnitTests {
         assertTrue(eigenPod.restakingDisabled(), "should disable");
         assertEq(delegationManagerMock.clearQueuedWithdrawalsCallCount(), 1, "cleanup should be called once");
         assertEq(delegationManagerMock.lastClearedDisabledPod(), address(this), "cleanup called for pod owner");
-        assertEq(
-            delegationManagerMock.getQueuedWithdrawalRoots(address(this)).length, 0, "queue entry should be cleared"
-        );
+        assertEq(delegationManagerMock.getQueuedWithdrawalRoots(address(this)).length, 0, "queue entry should be cleared");
     }
 
     function test_permanentlyDisableRestaking_revert_mixedWithdrawal() public {
@@ -2494,10 +2488,10 @@ contract EigenPodUnitTests_DisableRestaking is EigenPodUnitTests {
         // withdrawal so disable's net-restaked-vs-entitlement check passes (mirrors production: this
         // REL exists only because shares were queued out, which created the entitlement).
         IEigenPodTypes.Checkpoint memory c = pod.currentCheckpoint();
-        int256 netGwei =
-            int256(uint256(pod.withdrawableRestakedExecutionLayerGwei())) + int256(uint256(c.prevBeaconBalanceGwei)) + int256(c.balanceDeltasGwei);
+        int netGwei =
+            int(uint(pod.withdrawableRestakedExecutionLayerGwei())) + int(uint(c.prevBeaconBalanceGwei)) + int(c.balanceDeltasGwei);
         eigenPodManagerMock.setPodOwnerShares(podOwner, 0);
-        _registerCoveringWithdrawal(podOwner, uint64(uint256(netGwei)));
+        _registerCoveringWithdrawal(podOwner, uint64(uint(netGwei)));
         _disablePod(pod);
 
         // Disabling zeroes REL: the formerly-reserved ETH is now treated as non-restaked.
@@ -2548,7 +2542,7 @@ contract EigenPodUnitTests_DisableRestaking is EigenPodUnitTests {
         _wireDisablePreconditions();
 
         cheats.prank(pauser);
-        eigenPodManagerMock.pause(2 ** PAUSED_PERMANENTLY_DISABLE_RESTAKING);
+        eigenPodManagerMock.pause(2**PAUSED_PERMANENTLY_DISABLE_RESTAKING);
 
         cheats.expectRevert(IEigenPodErrors.CurrentlyPaused.selector);
         eigenPod.permanentlyDisableRestaking();
@@ -2561,7 +2555,7 @@ contract EigenPodUnitTests_DisableRestaking is EigenPodUnitTests {
         _seedPodWithETH(1 ether);
 
         cheats.prank(pauser);
-        eigenPodManagerMock.pause(2 ** PAUSED_PERMANENTLY_DISABLE_RESTAKING);
+        eigenPodManagerMock.pause(2**PAUSED_PERMANENTLY_DISABLE_RESTAKING);
 
         address recipient = cheats.addr(0xBEEF);
         cheats.expectRevert(IEigenPodErrors.CurrentlyPaused.selector);
